Set up a container-assisted development environment
There is now a development Containerfile, and a makefile for setting
it up.  Just running `make` will set up and run the containers in the
background.  If you run `make` again, it will re-start the containers.

There are two containers, with some shared volumes.  The main container
is `worldmaster-development-runserver`, which runs migrations and runs
`./manage.py runserver`.

There is also a `worldmaster-development-tsc` container, which watches
the `./tsc` directory for changes and re-runs `tsc` every time one of
them changes.

For the most part, this means that you can just run the containers and
develop with them running in the background.  You'll get hot-reloading
and automatic TypeScript recompilation.  You might need to run `make`
or `make worldmaster-development-runserver` to re-run migrations if you
make more migrations.

`make clean` will destroy all containers and their volumes.  You can use
that if you get your stuff stuck.

This can still be enhanced.  Other possibilities in the future:

* Another container that is not in the background for generating
  migrations.

* Either a container for `./manage.py shell_plus` or an exec of the
  runserver container.

* A development postgres container, if we want to lean on postgres-
  specific features.

* Production server containers.

* More help for containers in the Makefile, such as individual volume or
  container cleaning.