small changes to justfile and a note in the Containerfile
Justfile development recipe: Allow ports to be specified as either a
single port or a pair.

development.Containerfile:  Add a comment explaining that bind mounts
seem to map the local uid and gid outside the container to root inside.
The understanding I take is that development containers usually should
run with the root user, so they can make changes they might need to to
the bind-mounted source directory, or any other bind-mounted directory.