add roles, remove multi-table inheritance
The default django permissions system only appears to work with Model
types out of the box, not individual objects.  Thus, we have introduced
our own Role type for managing roles.

MASTER is applied to the user who creates a world.

EDITOR is applied to the user who creates any entities other than a world.

This is because a user may create an entity like a magic object, for
instance, and the DM may put notes into it that the player can't see
yet.  Players may also give other users edit permission on things that
they create.  If they had MASTER privileges, no sections under the
things they create can be hidden from them.

Note also that as a result, a player can not delete things that they
create, because they could delete things that they can't even see.

This is not a perfect system, and the implementation is a little wonky
in parts (for instance, if you want to edit a section, it needs to be
checked if you have EDITOR or MASTER on the section or MASTER on the
world, but the section's wiki backlink doesn't have a ref to the world
itself, so the world needs to be passed in separately).

Multi-table inheritance is a little bit kludgy and hard to visualize.
Very problematically, it also sets up the foreign keys to be primary
keys, so it would be extremely difficult to change later.  This would be
fine if we were setting up a straight-up extension class, but our models
are too involved to couple them that tightly.

As such, we have removed all multi-table inheritance.  We've added some
abstract bases and signals to make management of this all easier.

The migrations had to be wiped and rebuilt to account for the PK change.