simplify containers, removing unnecessary volumes
Containers now interact with the host via bind mounts directly for most
needs.  The UID and GID are the same in the container and out (really,
this is mapped through a subuid and subgid that aren't the same, but it
works either way).

The development db is put into a `dev` directory so that can be writably
bind mounted.

Migration directories are properly mounted for the same reason.