add players, fix broken signals
The players junction table is added now.  This allows arbitrary users
to be associated with arbitrary worlds.  The players automatically get
VIEWER on the world.

Some of the role management was broken by the last change, so this fixes
that as well.  We need some basic sets of automatic tests as soon as we
can get them.

Signals are now re-worked slightly.  They already were recursive, but
now they are automatically recursive by other signals.  This makes some
things more robust, and also accounts for things like adding a recursive
signal.