roles: implicit and inherited roles resolved dynamically; add tests
Previously, all implicit and inherited roles were deduced with signals
and actually stored into the database.  This had the implication of a
lot of extra rows in the database.  If somebody had a master role in a
world (a common case), they would automatically have master, editor, and
viewer in every single object in that world.

Now we use a recursive CTE, stored in a view and invoked via a non-
managed model.  This is significantly better.  Due to less index
contention, it might even perform better, and definitely uses much less
storage.

In order to validate these changes, a test suite has also been added.