add natural key stuff, reduce code repetition, add ResolvedRole CreateModel, move roletarget methods to the manager, add test for deduplicated CTE
Natural keys are sharing methods where they can now.

ResolvedRole is a managed = False model, but these still want to
generate a CreateModel migration, so that's been merged in.

According to [the django documentation](https://docs.djangoproject.com/en/4.2/topics/db/managers/#custom-managers):

> Adding extra Manager methods is the preferred way to add “table-level”
> functionality to your models. (For “row-level” functionality - i.e.,
> functions that act on a single instance of a model object - use Model
> methods, not custom Manager methods.)

So because these methods are table-level, providing filters, they are
moved to the manager. At some point, it might make sense to make them
work on QuerySets instead, so they can be applied later in a QuerySet
chain, but this is good enough for now.

The CTE provided duplicated roles before.  This has been corrected. At
some point, a materialized CTE might be better, or the CTE might need to
be made much more efficient somehow.